Move marshalling methods for Type to partial file
